refactor: move areStocksDistinct to utils
