+++ src/lib/utils/medication-stock.ts
@@ -20,3 +20,22 @@ export type MedicationStockWithBrandAndActiveIngredient = MedicationStock & {
   medicationBrandName: string;
   medicationActiveIngredientName: string;
 };
+
+type Payload = {
+  location: string;
+  stockStatus: StockStatus;
+  quantity: number;
+  remarks: string | undefined;
+};
+
+export function areStocksDistinct(stocks: Payload[]) {
+  const stocksSet = new Set();
+  for (let i = 0; i < stocks.length; i++) {
+    const stringified = JSON.stringify(stocks[i]);
+    if (stocksSet.has(stringified)) {
+      return false;
+    }
+    stocksSet.add(stringified);
+  }
+  return true;
+}
